Feed Respoke auth token to logged in html
This patch changes the initial logged in html such that it will
have a token as soon as the page is rendered. This allows the Respoke
client in the browser to immediately connect when the page loads,
instead of having to make a round-trip to the server to fetch a token
first. To show how to use this, I also added a very simple app.js to
demonstrate connecting using the token from the html and then reconnecting
using the API to fetch a fresh token.

Also as a part of this patch, I refactored the RespokeService to accept
the ttl and roleId on the constructor, instead of requiring you to pass
it each time you request an auth token.

Closes #4